requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.5.0
seaborn>=0.11.0
reportlab>=4.0.0
weasyprint>=60.0
markdown>=3.5.0
